fix: ID sequence generator fix

+++ Common/src/ClientCommands/Update.java
@@ -10,23 +10,23 @@ import Utils.ResponseCode;
 
 public class Update implements ClientCommand {
 
-    private long id;
+    private int id;
     private Dragon dragon;
 
     public Dragon getDragon() {
         return dragon;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
     @Override
     public void prepareRequest(String[] args) throws WrongArgument, NotEnoughArgs {
         if(args.length < 2) throw new NotEnoughArgs("Command requires \"id\" argument");
-        long id;
+        int id;
         try{
-            id = Long.parseLong(args[1]);
+            id = Integer.parseInt(args[1]);
             this.id = id;
         }
         catch (NumberFormatException e){
